refactor: split cow pool module
